refactor: Rename PartieDeChasseBuilder namespace for better organization
